refactor(parquet-datasource): extract ExpectedPruning into a shared test_util module
`ExpectedPruning` is the only test helper shared between the row-group
statistics tests and the bloom-filter tests. Move it into a new
`#[cfg(test)] mod test_util` so the bloom-filter tests can be relocated
into `bloom_filter.rs` without it.

Its `assert` method previously reached into the private `access_plan`
field of `RowGroupAccessPlanFilter`; add a test-only `pub(crate)`
`access_plan()` accessor so the helper works from a sibling module
without widening the field's visibility.
